requirements.txt
altair==6.1.0
anyio==4.13.0
asn1crypto==1.5.1
attrs==26.1.0
blinker==1.9.0
boto3==1.43.1
botocore==1.43.1
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
filelock==3.29.0
gitdb==4.0.12
GitPython==3.1.49
h11==0.16.0
httptools==0.7.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
protobuf==7.34.1
pyarrow==24.0.0
pycparser==3.0
pydeck==0.9.2
PyJWT==2.12.1
pyOpenSSL==26.1.0
python-dateutil==2.9.0.post0
python-multipart==0.0.27
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
s3transfer==0.17.0
six==1.17.0
smmap==5.0.3
snowflake-connector-python==4.4.0
sortedcontainers==2.4.0
starlette==1.0.0
streamlit==1.57.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
websockets==16.0
